+++ virtualFittingApp/src/entities/coupon/api/coupon.api.ts
@@ -38,9 +38,20 @@ export const downloadCoupon = async (params: {
   const { campaignId, authUserId } = params;
   const requestBody: DownloadCouponRequestBody = { campaignId };
   
-  const response = await API_BASILIUM.post<DownloadCouponApiResponse>(`/b1/coupons/wallets`, requestBody, {
-    params: { authUserId }
-  });
-  console.log(`[쿠폰 다운로드 (ID: ${campaignId})] API 응답 성공:`, response.data);
-  return response.data.data;
+  try {
+    const response = await API_BASILIUM.post<DownloadCouponApiResponse>(
+      `/b1/coupons/wallets`, 
+      requestBody, 
+      {
+        params: { authUserId }
+      }
+    );
+    
+    console.log(`[쿠폰 다운로드 (ID: ${campaignId})] API 응답 성공:`, response.data);
+    return response.data.data;
+
+  } catch (error) {
+    console.error(`[쿠폰 다운로드 (ID: ${campaignId})] API 호출 실패:`, error);
+    return null; 
+  }
 };
