+++ virtualFittingApp/src/entities/coupon/index.ts
@@ -1,1 +1,3 @@
+export * from "./api/coupon.api";
+export * from "./api/coupon.keys";
 export * from "./model/types";
