+++ virtualFittingApp/src/features/coupon/index.ts
@@ -1,3 +1,2 @@
 export * from "./hooks/use-product-coupon"
 export * from "./ui/product-coupon";
-export * from "./coupon.keys";
